Upgrade to SDK 1.2.0
There are a few changes in here:

1. Bump the version in the daml.yaml.
2. Bump the version in the pom.xml.
3. Replace the shitty code for finding the main package id via the
   package service by `daml damlc inspect-dar` which is a much more
   sensible solution. This also removes the dependency on
   daml-lf-archive and any DAML-LF parsing.
4. Update all group ids of SDK dependencies from com.digitalasset to
   com.daml. I’ve left the ones for the project intact since I’m lazy
   and there is not much to be gained by changing them.
5. Remove the code for the old time model.
6. Switch from decimal to numeric (on the client side, the models have
   not been modified).
7. Update protobuf-java to match the upgrade on the SDK side.

I’ve run `make start` and the output looks roughly like the
screenshot. I believe the differences can be explained by concurrency
but it could also just be that the screenshot is severly outdated. I
honestly don’t understand the example enough to judge whether there
are other issues.